Don't try to parse response if previous parse needed more bytes
In the case of a large response (almost always due to a literal string
containing a message with attachments), we would attempt to parse the
available bytes every time new bytes came in. Instead, if the parser
knows it needs a certain number of bytes, only attempt to parse again
once enough bytes are available in the buffer.