Implement general linting strategy
We will suppress inappropriate lints locally, but deny by default.

Do not warn on missing docs yet

Do not make exception for the box-pointers lint

`box_pointers` is a lint that prevents use of owned heap memory. This
restriction seems unneeded for this library, but it exists as a lint.
Due to a bug in rustc it is not yet included in `#![deny(warnings)]`.
Remove the exception for now.