Adding a `BodyStructParser`
Regarding #92, I propose to add an helper parser to be able to search
out specific component from a `BodyStructure`. This will allow for
example to retrieve the body section ID of some attachment from the
BODYSTRUCTURE and perform a FETCH request in a second time to download
only this specific element.