Allow trailing whitespace in search responses.
This relates to #34.

While working on a project, I tried to connect to Yahoo's IMAP server,
and all SEARCH responses seemed to have a trailing whitespace. For
example in the case below, there is a whitespace (even though it might
not seem obvious) right after the last id:

```
C: a5 UID SEARCH ALL
S: * SEARCH 169654 169655 169656 169657 169658 169659 169660 169661
169662 169663
```

I also happened to encounter this on empty SEARCH responses, which
led me to the issue linked above.