Add name attributes (#134)
Any programs that use the name attributes from the LIST command
currently have to match on strings. This change set adds a
`NameAttribute` enumeration and parses the name attributes so you can
match on the enumeration instead.

More values will be added to `NameAttribute` as more extensions get support
(e.g. RFC 6154 Special-Use Mailboxes). It should therefore be marked as
`non_exhaustive` to indicate that.