Add special-use mailboxes name attributes
This change only adds the name attributes from RFC 6154 Special-Use
Mailboxes and not any of the other features.

It should now be possible in a client app to identify the spam folder
without matching on the name of the mailbox. Matching on the name of the
mailbox is not reliable because it might be localized to the user's
language.